+++ Source/Rendering/Renderer.cpp
@@ -496,54 +496,6 @@ void MT::Renderer::RenderCopyPart(const Rect& rect, const Rect& source, const Te
     currentIndex += currentSize;
 }
 
-
-void MT::Renderer::RenderCopyEX(const Rect& rect, const Rect& source, const Texture* texture, const bool flip, const float rotation) {
-    if (!texture) { return; }
-    if (currentTexture != texture->texture) {
-        Present(false);
-        glBindTexture(GL_TEXTURE_2D, texture->texture);
-        currentTexture = texture->texture;
-    }
-
-    if (currentProgram != renderCopyExId) {
-        Present(false);
-        glBindVertexArray(copyExVao);
-        currentProgram = renderCopyExId;
-        glUseProgram(currentProgram);
-    }
-    currentSize = renderCopyExSize;
-
-    const float sourceX = static_cast<float>(source.x) / texture->w;
-    const float sourceY = static_cast<float>(source.y) / texture->h;
-    const float sourceW = static_cast<float>(source.w) / texture->w;
-    const float sourceH = static_cast<float>(source.h) / texture->h;
-
-    if (currentIndex + currentSize > Renderer::batchSize) {
-        Present(false);
-    }
-
-    float* ptr = &globalVertices[currentIndex];
-    ptr[0] = static_cast<float>(rect.x);
-    ptr[1] = static_cast<float>(rect.y);
-    ptr[2] = static_cast<float>(rect.w);
-    ptr[3] = static_cast<float>(rect.h);
-    if (flip) {
-        ptr[4] = sourceX + sourceW; 
-        ptr[6] = -sourceW;          
-    }
-    else {
-        ptr[4] = sourceX;      
-        ptr[6] = sourceW;
-    }
-    ptr[5] = sourceY;
-    ptr[7] = sourceH;
-    ptr[8] = texture->alpha;
-    ptr[9] = rotation;
-
-    currentIndex += currentSize;
-}
-
-
 void MT::Renderer::RenderCopyCircle(const Rect& rect, const Texture* texture, const float radius) {
     if (!texture) { return; }
     if (!vievPort.IsColliding(rect)) {return;}
@@ -684,58 +636,6 @@ void MT::Renderer::RenderCopyRounded(const Rect& rect, const Texture* texture, i
     ptr[3] = static_cast<float>(rect.h);
     ptr[4] = texture->alpha;
     ptr[5] = static_cast<float>(roundingSize);
-
-    currentIndex += currentSize;
-}
-
-
-void MT::Renderer::RenderCopyFiltered(const Rect& rect, const Rect& source, const Texture* texture, const Color& filter) {
-    if (!texture) { return; }
-    if (!vievPort.IsColliding(rect)) {return;}
-
-    if (currentProgram != renderCopyFilterId) {
-        Present(false);
-        glBindVertexArray(filteredVao);
-        currentProgram = renderCopyFilterId;
-        glUseProgram(renderCopyFilterId);
-    }
-
-    if (currentTexture != texture->texture) {
-        Present(false);
-        glBindTexture(GL_TEXTURE_2D, texture->texture);
-        currentTexture = texture->texture;
-    }
-
-
-    uint16_t iRG = filter.R;
-    iRG <<= 8;
-    iRG += filter.G;
-    uint16_t iBA = filter.B;
-    iBA <<= 8;
-    iBA += static_cast<unsigned char>(texture->alpha * 255);
-
-    currentSize = renderFilteredSize;
-
-    if (currentIndex + currentSize > Renderer::batchSize) {
-        Present(false);
-    }
-
-    const float sourceX = static_cast<float>(source.x) / texture->w;
-    const float sourceY = static_cast<float>(source.y) / texture->h;
-    const float sourceW = static_cast<float>(source.w) / texture->w;
-    const float sourceH = static_cast<float>(source.h) / texture->h;
-
-    float* ptr = &globalVertices[currentIndex];
-    ptr[0] = static_cast<float>(rect.x);
-    ptr[1] = static_cast<float>(rect.y);
-    ptr[2] = static_cast<float>(rect.w);
-    ptr[3] = static_cast<float>(rect.h);
-    ptr[4] = sourceX;
-    ptr[5] = sourceY;
-    ptr[6] = sourceW;
-    ptr[7] = sourceH;
-    ptr[8] = static_cast<float>(iRG);
-    ptr[9] = static_cast<float>(iBA);
 
     currentIndex += currentSize;
 }
@@ -1065,47 +965,6 @@ void MT::Renderer::RenderCopyPartUPR(const Rect& rect, const Rect& source, const
     currentIndex += currentSize;
 }
 
-void MT::Renderer::RenderCopyEXUPR(const Rect& rect, const Rect& source, const Texture* texture, const bool flip, const float rotation) {
-    if (!texture) { return; }
-    if (!vievPort.IsColliding(rect)) { return; }
-
-    CheckUPRProgram();
-
-    if (currentTexture != texture->texture) {
-        Present(false);
-        glBindTexture(GL_TEXTURE_2D, texture->texture);
-        currentTexture = texture->texture;
-    }
-
-
-    const float sourceX = static_cast<float>(source.x) / texture->w;
-    const float sourceY = static_cast<float>(source.y) / texture->h;
-    const float sourceW = static_cast<float>(source.w) / texture->w;
-    const float sourceH = static_cast<float>(source.h) / texture->h;
-
-    float* ptr = &globalVertices[currentIndex];
-    ptr[0] = static_cast<float>(rect.x);
-    ptr[1] = static_cast<float>(rect.y);
-    ptr[2] = static_cast<float>(rect.w);
-    ptr[3] = static_cast<float>(rect.h);
-
-    if (flip) {
-        ptr[4] = sourceX + sourceW;
-        ptr[6] = -sourceW;
-    }
-    else {
-        ptr[4] = sourceX;
-        ptr[6] = sourceW;
-    }
-    ptr[5] = sourceY;
-    ptr[7] = sourceH;
-    ptr[8] = static_cast<float>(texture->alpha); ptr[9] = rotation; ptr[10] = 0.0f; ptr[11] = 0.0f;
-    ptr[12] = 0.0f;
-    ptr[13] = 4.0f; //ShaderID
-
-    currentIndex += currentSize;
-}
-
 void MT::Renderer::RenderCopyCircleUPR(const Rect& rect, const Texture* texture, const float radius) {
     if (!texture) { return; }
     if (!vievPort.IsColliding(rect)) { return; }
@@ -1200,42 +1059,6 @@ void MT::Renderer::RenderCopyRoundedUPR(const MT::Rect& rect, const MT::Texture*
     ptr[8] = 0.0f; ptr[9] = 0.0f; ptr[10] = 0.0f; ptr[11] = 0.0f;
     ptr[12] = 0.0f;
     ptr[13] = 8.0f; //ShaderID
-
-    currentIndex += currentSize;
-}
-
-void MT::Renderer::RenderCopyFilteredUPR(const Rect& rect, const Rect& source, const Texture* texture, const Color& filter) {
-    if (!texture) { return; }
-    if (!vievPort.IsColliding(rect)) { return; }
-
-    CheckUPRProgram();
-
-    if (currentTexture != texture->texture) {
-        Present(false);
-        glBindTexture(GL_TEXTURE_2D, texture->texture);
-        currentTexture = texture->texture;
-    }
-
-
-    const float sourceX = static_cast<float>(source.x) / texture->w;
-    const float sourceY = static_cast<float>(source.y) / texture->h;
-    const float sourceW = static_cast<float>(source.w) / texture->w;
-    const float sourceH = static_cast<float>(source.h) / texture->h;
-
-    float* ptr = &globalVertices[currentIndex];
-    ptr[0] = static_cast<float>(rect.x);
-    ptr[1] = static_cast<float>(rect.y);
-    ptr[2] = static_cast<float>(rect.w);
-    ptr[3] = static_cast<float>(rect.h);
-
-    ptr[4] = sourceX; ptr[5] = sourceY; ptr[6] = sourceW; ptr[7] = sourceH;
-    ptr[8] = static_cast<float>(filter.R); 
-    ptr[9] = static_cast<float>(filter.G);
-    ptr[10] = static_cast<float>(filter.B);
-    ptr[11] = texture->alpha;
-
-    ptr[12] = 0.0f;
-    ptr[13] = 9.0f; //ShaderID
 
     currentIndex += currentSize;
 }