+++ CHANGELOG.md
@@ -186,3 +186,4 @@ was removed or the new one was added to folder
 
 2.03
 - New optional logger class
+- RenderCopyEx and RenderCopyFiltered without sourceRectangle and when texture is nullptr will no longer cause seg fault